Some more clean-up.

+++ src/devices/ib/GPI2_IB_GRP.c
@@ -105,7 +105,7 @@ pgaspi_dev_post_write(void *local_addr, int length, int dst, void *remote_addr, 
   swr.send_flags = (length == 1) ? (IBV_SEND_SIGNALED | IBV_SEND_INLINE) : IBV_SEND_SIGNALED;
   swr.next = NULL;
 
-  swr.wr.rdma.remote_addr = remote_addr;
+  swr.wr.rdma.remote_addr = (uint64_t) remote_addr;
   swr.wr.rdma.rkey = glb_gaspi_group_ib[group].rrcd[dst].rkeyGroup;
   swr.wr_id = dst;
